refactor: 입력 값의 유효성을 검사하는 함수들을 Validator 클래스로 이전
BaseballGame 클래스에 있던 유효성 검사 함수들(checkIsValidInput, checkLength, checkIsAllValidNumber, checkDuplicatedNumber)를
Validator 클래스로 이동
